feat: Add comprehensive README, model files, and complete backend/frontend
- Add detailed README.md with MIMIC-IV dataset instructions
- Add .gitignore for sensitive data (dataset, db, venv)
- Add .gitattributes for Git LFS (large model files)
- Add new_model/ with trained XGBoost model artifacts
- Add SQL query for MIMIC-IV data extraction
- Update backend with prediction endpoints
- Update frontend with modern dashboard UI

+++ backend/schemas.py
@@ -139,7 +139,7 @@ class PredictionInput(BaseModel):
     ntprobnp_max: Optional[float] = None
 
 class PredictionOutput(BaseModel):
-    sepsis: int
+    sepsis: float
     respiration: float
     coagulation: float
     liver: float
